backend/requirements.txt
fastapi
uvicorn
feedparser
python-dotenv
langgraph
httpx
groq

# v0.7.0: LLM-powered aggregation, Groq, LangGraph
